requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
motor==3.7.0
redis==5.2.1
celery==5.4.0
openai==1.63.2
python-dotenv==1.0.1
httpx==0.28.1
python-multipart==0.0.20
